+++ ci/test_rasters.py
@@ -25,7 +25,7 @@ import rasterio
 import shapely
 import xarray as xr
 
-from CI.SCRIPTS.script_utils import KAPUT_KWARGS, dask_env, rasters_path, s3_env
+from ci.script_utils import KAPUT_KWARGS, dask_env, rasters_path, s3_env
 from sertit import ci, path, rasters, vectors
 from sertit.rasters import (
     FLOAT_NODATA,