+++ ci/test_rasters_rio.py
@@ -24,7 +24,7 @@ import rasterio
 import shapely
 from rasterio.windows import Window
 
-from CI.SCRIPTS.script_utils import KAPUT_KWARGS, rasters_path, s3_env
+from ci.script_utils import KAPUT_KWARGS, rasters_path, s3_env
 from sertit import ci, rasters_rio, vectors
 from sertit.vectors import EPSG_4326