+++ sertit/rasters.py
@@ -1126,7 +1126,10 @@ def write(
 
     # Write COGs
     is_written = False
+    blocksize = None
     if is_cog:
+        blocksize = 128 if (xds.rio.height < 1000 or xds.rio.width < 1000) else None
+
         if write_cogs_with_dask:
             try:
                 from dask import optimize
@@ -1145,6 +1148,7 @@ def write(
                     ),
                     str(output_path),
                     stats=compute_stats,
+                    blocksize=blocksize,
                 )
 
                 (delayed,) = optimize(delayed)
@@ -1167,6 +1171,8 @@ def write(
                 "Loading raster in memory as COG has been asked not to be written by Dask. Be careful about MemoryErrors!"
             )
             xds = xds.load()
+            if blocksize is not None:
+                kwargs["BLOCKSIZE"] = blocksize
 
         if not is_written:
             # Write with windows as we don't want to_raster to blow up the RAM
@@ -1904,13 +1910,12 @@ def slope(
         AnyXrDataStructure: Slope
     """
     try:
-        from xarray.ufuncs import tan
         from xrspatial import slope
 
         xds = slope(xds)
 
         if in_pct:
-            xds = 100 * tan(xds * DEG_2_RAD)
+            xds = 100 * np.tan(xds * DEG_2_RAD)
         elif in_rad:
             xds = xds * DEG_2_RAD
     except ImportError: